+++ src/extension/connector/App.vue
@@ -1,37 +1,38 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import { useAppStore } from "@/stores/appStore";
-import { isPopup } from "@/common/browser";
-import KyaModal from "@/components/KYAModal.vue";
 import WalletLogo from "@/components/WalletLogo.vue";
-import WalletHeader from "@/components/WalletHeader.vue";
-import NavHeader from "@/components/NavHeader.vue";
+import WalletItem from "@/components/WalletItem.vue";
+import { useWalletStore } from "@/stores/walletStore";
 
 const app = useAppStore();
-const maxWidth = computed(() => (isPopup() ? "max-w-365px" : undefined));
+const wallet = useWalletStore();
+
+const currentWalletItem = computed(() => app.wallets.find((w) => w.id === wallet.id));
 </script>
 
 <template>
-  <div class="app" :class="maxWidth">
+  <div class="app">
     <div
-      v-if="$route.meta.fullPage"
-      class="flex flex-row p-4 gap-4 items-center justify-between bg-gray-100 border-b-1 border-gray-200"
+      class="flex flex-row py-4 px-6 gap-4 items-center justify-between bg-gray-100 border-b-1 border-gray-200"
     >
-      <wallet-logo root-class="ml-2" content-class="w-11 h-11" />
-      <h1 class="text-base font-semibold w-full pl-2">
+      <wallet-logo content-class="w-11 h-11" />
+      <wallet-item
+        v-if="!app.loading && !$route.meta.fullPage && currentWalletItem"
+        :key="wallet.id"
+        class="w-min"
+        :reverse="true"
+        :wallet="currentWalletItem"
+        :loading="wallet.loading || wallet.syncing"
+      />
+      <h1 v-else class="text-base font-semibold w-full">
         <template v-if="$route.meta.title">{{ $route.meta.title }}</template>
         <template v-else>Nautilus Wallet</template>
       </h1>
     </div>
-    <template v-else>
-      <wallet-header />
-      <nav-header v-if="$route.query.popup !== 'true'" />
-    </template>
 
     <div class="flex-grow overflow-y-auto overflow-x-hidden p-4">
       <router-view />
     </div>
-
-    <kya-modal :active="!app.loading && !app.settings.isKyaAccepted" />
   </div>
 </template>
