+++ src/extension/connector/main.ts
@@ -1,8 +1,8 @@
 import { createApp } from "vue";
 import App from "./App.vue";
 import { router } from "./router";
+import { registerRpcHooks } from "./rpc/uiRpcHandlers";
 import { setupApp } from "@/boot/commonSetup";
-import { registerRpcHooks } from "@/rpc/uiRpcHandlers";
 
 registerRpcHooks();
 setupApp(createApp(App)).use(router).mount("#app");
