+++ src/extension/connector/router.ts
@@ -1,32 +1,32 @@
 import { createRouter, createWebHashHistory, RouteRecordRaw } from "vue-router";
 
 // Views
-const connectView = () => import("@/views/connector/ConnectView.vue");
-const authView = () => import("@/views/connector/AuthView.vue");
-const signDataView = () => import("@/views/connector/SignDataView.vue");
-const signTxView = () => import("@/views/connector/SignTxConfirmView.vue");
+const connectView = () => import("./views/ConnectView.vue");
+const authView = () => import("./views/AuthView.vue");
+const signDataView = () => import("./views/SignDataView.vue");
+const signTxView = () => import("./views/SignTxConfirmView.vue");
 
 const routes: Array<RouteRecordRaw> = [
   {
-    path: "/connector/connect",
+    path: "/connect",
     name: "connector-connect",
     component: connectView,
     meta: { title: "Access request", fullPage: true }
   },
   {
-    path: "/connector/sign/tx",
+    path: "/sign/tx",
     name: "connector-sign-tx",
     component: signTxView,
     meta: { title: "Transaction signature" }
   },
   {
-    path: "/connector/sign/data",
+    path: "/sign/data",
     name: "connector-sign-data",
     component: signDataView,
     meta: { title: "Data signature" }
   },
   {
-    path: "/connector/auth",
+    path: "/auth",
     name: "connector-auth",
     component: authView,
     meta: { title: "Authorization" }
