+++ src/extension/connector/rpc/asyncRequestQueue.ts
@@ -1,5 +1,5 @@
 import type { AuthArgs, SignDataArgs, SignTxArgs, SignTxInputsArgs } from "@/types/d.ts/webext-rpc";
-import { InternalRequest } from "@/rpc/protocol";
+import { InternalRequest } from "@/extension/connector/rpc/protocol";
 
 export type AsyncRequest<T = unknown> = {
   type: InternalRequest;
