+++ src/extension/connector/views/ConnectView.vue
@@ -3,9 +3,9 @@ import { onMounted, ref } from "vue";
 import { useEventListener } from "@vueuse/core";
 import DappPlateHeader from "@/components/DappPlateHeader.vue";
 import WalletItem from "@/components/WalletItem.vue";
-import { AsyncRequest } from "@/rpc/asyncRequestQueue";
-import { queue } from "@/rpc/uiRpcHandlers";
-import { InternalRequest } from "@/rpc/protocol";
+import { AsyncRequest } from "@/extension/connector/rpc/asyncRequestQueue";
+import { queue } from "@/extension/connector/rpc/uiRpcHandlers";
+import { InternalRequest } from "@/extension/connector/rpc/protocol";
 import { connectedDAppsDbService } from "@/database/connectedDAppsDbService";
 import { IDbWallet } from "@/types/database";
 import { walletsDbService } from "@/database/walletsDbService";
