+++ src/extension/connector/views/SignTxConfirmView.vue
@@ -3,9 +3,9 @@ import { computed, onMounted, ref, watch } from "vue";
 import { SignedTransaction, some } from "@fleet-sdk/common";
 import { useEventListener } from "@vueuse/core";
 import TxSignView from "@/components/TxSignView.vue";
-import { AsyncRequest } from "@/rpc/asyncRequestQueue";
-import { error, InternalRequest, success } from "@/rpc/protocol";
-import { queue } from "@/rpc/uiRpcHandlers";
+import { AsyncRequest } from "@/extension/connector/rpc/asyncRequestQueue";
+import { error, InternalRequest, success } from "@/extension/connector/rpc/protocol";
+import { queue } from "@/extension/connector/rpc/uiRpcHandlers";
 import { SignTxArgs, SignTxInputsArgs } from "@/types/d.ts/webext-rpc";
 import { connectedDAppsDbService } from "@/database/connectedDAppsDbService";
 import { APIErrorCode, SignErrorCode } from "@/types/connector";
