+++ src/extension/content-scripts/contentScript.ts
@@ -1,5 +1,10 @@
 import { allowWindowMessaging, onMessage, sendMessage } from "webext-bridge/content-script";
-import { buildNamespaceFor, ExternalEvent, ExternalRequest, InternalRequest } from "@/rpc/protocol";
+import {
+  buildNamespaceFor,
+  ExternalEvent,
+  ExternalRequest,
+  InternalRequest
+} from "../connector/rpc/protocol";
 import { EXT_ENTRY_ROOT } from "@/constants/extension";
 
 const CONSOLE_PREFIX = "[Nautilus]";
