+++ src/extension/content-scripts/injected.ts
@@ -2,7 +2,12 @@ import { sendMessage, setNamespace } from "webext-bridge/window";
 import { SelectionTarget } from "@nautilus-js/eip12-types";
 import type { EIP12UnsignedTransaction, SignedTransaction } from "@fleet-sdk/common";
 import type { JsonValue } from "type-fest";
-import { buildNamespaceFor, ExternalEvent, ExternalRequest, Result } from "@/rpc/protocol";
+import {
+  buildNamespaceFor,
+  ExternalEvent,
+  ExternalRequest,
+  Result
+} from "../connector/rpc/protocol";
 import { APIErrorCode } from "@/types/connector";
 
 const CONTENT_SCRIPT = "content-script";
